chore: move GitHub pagination test into github_test.go
